Improve toolbar content
